debug: ts compiler failing test

+++ packages/service/test/cli/compile.test.mjs
@@ -164,7 +164,11 @@ t.test('start command should not compile typescript plugin with errors', async (
     await childProcess
     t.fail('should not compile bad typescript plugin')
   } catch (err) {
-    t.equal(err.stderr.includes('Found 1 error'), true)
+    if (!err.stderr.includes('Found 1 error')) {
+      t.comment(err.stdout)
+      t.comment(err.stderr)
+      t.fail('should throw one ts error')
+    }
     childProcess.kill('SIGINT')
   }
